Move external codegen test helpers into utils (#9008)
This is so they can be re-used as part of other tests which don't extend test_external_codegen.py

I've identified that `test_json_runtime.py` and `test_pass_partition_graph.py` use a very similar but slightly different variant of these functions for future iterations.
